requirements.txt
allure-python-commons~=2.13.5
selenium~=4.26.1
webdriver-manager~=4.0.2
requests~=2.32.3